float to double, plus a timeCat

+++ src/Sound/Tidal/Params.hs
@@ -4,7 +4,7 @@ import qualified Data.Map.Strict as Map
 
 import Sound.Tidal.Pattern
 
-cf :: String -> Pattern Float -> ControlPattern
+cf :: String -> Pattern Double -> ControlPattern
 cf name = fmap (Map.singleton name . VF)
 
 ci :: String -> Pattern Int -> ControlPattern
@@ -13,37 +13,37 @@ ci name = fmap (Map.singleton name . VI)
 cs :: String -> Pattern String -> ControlPattern
 cs name = fmap (Map.singleton name . VS)
 
-begin :: Pattern Float -> ControlPattern
+begin :: Pattern Double -> ControlPattern
 begin = cf "begin"
 
 cut :: Pattern Int -> ControlPattern
 cut = ci "cut"
 
-end :: Pattern Float -> ControlPattern
+end :: Pattern Double -> ControlPattern
 end = cf "end"
 
-gain :: Pattern Float -> ControlPattern
+gain :: Pattern Double -> ControlPattern
 gain = cf "end"
 
-loop :: Pattern Float -> ControlPattern
+loop :: Pattern Double -> ControlPattern
 loop = cf "loop"
 
-n :: Pattern Float -> ControlPattern
+n :: Pattern Double -> ControlPattern
 n = cf "n"
 
-note :: Pattern Float -> ControlPattern
+note :: Pattern Double -> ControlPattern
 note = cf "note"
 
 s :: Pattern String -> ControlPattern
 s = cs "s"
 
-shape :: Pattern Float -> ControlPattern
+shape :: Pattern Double -> ControlPattern
 shape = cf "shape"
 
-speed :: Pattern Float -> ControlPattern
+speed :: Pattern Double -> ControlPattern
 speed = cf "speed"
 
-pan :: Pattern Float -> ControlPattern
+pan :: Pattern Double -> ControlPattern
 pan = cf "pan"
 
 unit :: Pattern String -> ControlPattern
